add depth_first_fold to the interface
The necessity for this function comes from a future improvement
requiring the accumulation of a value along paths in the graphs.

Since depth_first_fold and breadth_first_fold should behave in a similar way
this commit creates an uniformisation through a configuration structure.

In the case of the depth first path it is clear how the local_value
should be accumulated. For the breadth first path, however, the choice
was made to start again from the initial value at each level of
recursion.

This reveal a difference in behaviour, indeed the depth first path the
value changes even though it is not shared by all paths whereas the
breadth first path the value is constantly reinitialized.