add a tool to generate the data for the paper
The tool is compiled by calling "make filter".

The takes the path to a file containing pairs of reference to
allocation and release functions, and the path to the root of the
relative address used in the references.

It returns a filtered list, filtering away the pairs that do not verify:
allocation is not in error handling code
release is in error handling code