improve performance by reducing the number of assignements considered
Every assignment was considered but this patch reduces the number of
assignement to the one involving variables that are returned.

This changes greatly the output of tests since we are not considering
the function is not returning all error assignements anymore.

It may be interesting to investigate a version of the agorithm where
we look exclusively for the last assignments before the return.