add the last traversed edge to the configuration of the fold
The previous logic only allowed to reason about nodes, this patch adds
the possibility to reason about the edge that was used to get to that
node.

The data structure representing a node is now a record rather than a
couple, and the edge has been added as the second member of a couple.

It would be worth introducing a record rather than using couple.