changes the edge data structure to only include the index
The edge data structure used to include the complute nodes for the
initial node and the destination node. Because at the time the edge are
created the nodes are not yet annotated this lead to an error later on
when accessing them. This patch remove the reference to the node
structure, the user now has to use the KeyMap to get the structure.