fix the alias analysis and properly ignore the error branch handling a given allocation
The former implementation only made sure that the last reference of the
resource before a fault block was not an allocation. This patch makes sure that
the given block is not handling an error during the resouce allocation.

This first modification revealed that the alias analysis in
get_candidate_blocks was not initialized correctly. Therefore this patch
fixes that initialisation.