fix undetected array access
Hector did not handle the cases where the name of the index varied
within a function. The commit make the hypothesis that most of the time,
array access are mdone wihtin a loop and therefore we can ignore the
actual identifier being used.

A more robust version would ensure to be within a for loop and for the
index to be the loop variable.